+++ package.json
@@ -1,11 +1,8 @@
 {
-  "name": "client",
+  "name": "moodmuse",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "eslint-config-prettier": "^6.5.0",
-    "eslint-plugin-prettier": "^3.1.1",
-    "prettier": "^1.19.1",
     "react": "^16.11.0",
     "react-dom": "^16.11.0",
     "react-scripts": "3.2.0",
@@ -32,12 +29,15 @@
   "devDependencies": {
     "eslint": "6.1.0",
     "eslint-config-airbnb": "18.0.1",
+    "eslint-config-prettier": "^6.5.0",
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-prettier": "^3.1.1",
     "eslint-plugin-react": "^7.14.3",
     "eslint-plugin-react-hooks": "^1.7.0",
     "husky": ">=1",
-    "lint-staged": ">=8"
+    "lint-staged": ">=8",
+    "prettier": "^1.19.1"
   },
   "husky": {
     "hooks": {
